requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
aioredis==2.0.1
